requirements.txt
backoff==2.2.1
censys==2.2.0
certifi==2026.2.25
charset-normalizer==3.4.7
cssselect==1.4.0
dnspython==2.8.0
html-similarity==0.3.2
idna==3.11
ipaddress==1.0.19
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
parsel==1.2.0
Pygments==2.20.0
requests==2.31.0
rich==14.3.3
six==1.17.0
urllib3==1.26.18
w3lib==2.4.1
